Commit 11-19-2013
Cleaned up a lot of the code.

Added to the Auth system so that when a new user registers that it auto
creates a user profile record for them.

Added the profile dashboard
Added the profile manage method
Added the profile edit method
Added the profile success view
Added the profile edit form - still have to code it up