Commit 11-24-2013
Optinized the Template module to use just one method instead of having
multiple methods for each view layout, SEE: the template controller.

FIXED the comment controller was missing the extension .php
ADDED The manage methods and views to the 0 - Copy New Module. This way
we only have to edit the methods when creating a new module.