Commit 11-28-2013
ADDED: The new Auth Login and Registration form.
UPDATED: The Auth to use a single login method for both login and
registration. Still needs a little more work, but it is working.
ADDED: to the Home page breadcrumbs on the right a Welcome message for
user still needs to be coded.
CLEANED UP: A lot of garbage code laying around.