Commit 12-11-2013
RENAMED: The util_helper to functions_helper
POSTS: You should be able to login as admin and start adding posts! So
that we have something to work with.
Started coding the menu system.
ADD: A fontend Posts to the sidebar menu for displaying the posts, we
need sample data to work with!